refactor, fix($CarNameRepository): refactor for should_throwException_when_isNotIncludeComma test case
- change variable, field, class naming:
  carNames -> CarNameRepository
  carNames -> nameList
  list -> stringNameList

- add toString() method for test case
